Refactor test for msgmerge with compendium-overwrite option
 * Implemented a helper function `_get_expected` to standardize the expected PO file structure.
 * Renamed the option `c-overwrite` to `compendium-overwrite`

+++ tests/messages/test_frontend.py
@@ -750,25 +750,9 @@ class ConcatanationMessagesTestCase(unittest.TestCase):
             if os.path.isfile(file):
                     os.unlink(file)
 
-    def test_no_input_files(self):
-        with pytest.raises(OptionError):
-            self.cmd.finalize_options()
-
-    def test_no_output_file(self):
-        self.cmd.input_files = ['project/i18n/messages.pot']
-        with pytest.raises(OptionError):
-            self.cmd.finalize_options()
-
-    @freeze_time("1994-11-11")
-    def test_default(self):
-        self.cmd.input_files = [self.temp1, self.temp2]
-        self.cmd.output_file = self.output_file
-
-        self.cmd.finalize_options()
-        self.cmd.run()
-
+    def _get_expected(self, messages):
         date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
+        return fr"""# Translations template for PROJECT.
 # Copyright (C) 1994 ORGANIZATION
 # This file is distributed under the same license as the PROJECT project.
 # FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
@@ -787,7 +771,26 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Generated-By: Babel {VERSION}\n"
 
-#: simple.py:1
+""" + messages
+
+    def test_no_input_files(self):
+        with pytest.raises(OptionError):
+            self.cmd.finalize_options()
+
+    def test_no_output_file(self):
+        self.cmd.input_files = ['project/i18n/messages.pot']
+        with pytest.raises(OptionError):
+            self.cmd.finalize_options()
+
+    @freeze_time("1994-11-11")
+    def test_default(self):
+        self.cmd.input_files = [self.temp1, self.temp2]
+        self.cmd.output_file = self.output_file
+
+        self.cmd.finalize_options()
+        self.cmd.run()
+
+        expected_content = self._get_expected(fr"""#: simple.py:1
 #, flag1000
 msgid "other1"
 msgstr "Other 1"
@@ -814,10 +817,10 @@ msgstr "Other 3"
 msgid "other4"
 msgstr "Other 4"
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
-                actual_content = f.read()
+            actual_content = f.read()
         assert expected_content == actual_content
 
     @freeze_time("1994-11-11")
@@ -829,27 +832,7 @@ msgstr "Other 4"
         self.cmd.finalize_options()
         self.cmd.run()
 
-        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
-# Copyright (C) 1994 ORGANIZATION
-# This file is distributed under the same license as the PROJECT project.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PROJECT VERSION\n"
-"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: {date}\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel {VERSION}\n"
-
-#: simple.py:1
+        expected_content = self._get_expected(fr"""#: simple.py:1
 #, flag1000
 msgid "other1"
 msgstr "Other 1"
@@ -866,7 +849,7 @@ msgstr "Other 3"
 msgid "other4"
 msgstr "Other 4"
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
                 actual_content = f.read()
@@ -889,27 +872,7 @@ msgstr "Other 4"
         self.cmd.finalize_options()
         self.cmd.run()
 
-        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
-# Copyright (C) 1994 ORGANIZATION
-# This file is distributed under the same license as the PROJECT project.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PROJECT VERSION\n"
-"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: {date}\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel {VERSION}\n"
-
-#: hard.py:100 simple.py:100
+        expected_content = self._get_expected(fr"""#: hard.py:100 simple.py:100
 #, flag1, flag1.2, flag4
 msgid "same"
 msgstr "Same"
@@ -919,10 +882,10 @@ msgstr "Same"
 msgid "almost_same"
 msgstr "Almost same"
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
-                actual_content = f.read()
+            actual_content = f.read()
         assert expected_content == actual_content
 
 
@@ -964,9 +927,39 @@ class MergeMessagesTestCase(unittest.TestCase):
             pofile.write_po(file, catalog)
 
     def tearDown(self):
-        for file in [self.temp_def, self.temp_ref, self.compendium, self.output_file]:
-            if os.path.isfile(file):
+        for file in [
+            self.temp_def,
+            self.temp_def + '~',
+            self.temp_def + '.bac',
+            self.temp_ref,
+            self.compendium,
+            self.output_file
+        ]:
+            if os.path.exists(file) and os.path.isfile(file):
                     os.unlink(file)
+
+    def _get_expected(self, messages):
+        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
+        return fr"""# Translations template for PROJECT.
+# Copyright (C) 1994 ORGANIZATION
+# This file is distributed under the same license as the PROJECT project.
+# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
+#
+#, fuzzy
+msgid ""
+msgstr ""
+"Project-Id-Version: PROJECT VERSION\n"
+"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
+"POT-Creation-Date: {date}\n"
+"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
+"Language-Team: LANGUAGE <LL@li.org>\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=utf-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Generated-By: Babel {VERSION}\n"
+
+""" + messages
 
     def test_no_input_files(self):
         with pytest.raises(OptionError):
@@ -1001,27 +994,7 @@ class MergeMessagesTestCase(unittest.TestCase):
         self.cmd.finalize_options()
         self.cmd.run()
 
-        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
-# Copyright (C) 1994 ORGANIZATION
-# This file is distributed under the same license as the PROJECT project.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PROJECT VERSION\n"
-"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: {date}\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel {VERSION}\n"
-
-msgid "word1"
+        expected_content = self._get_expected(fr"""msgid "word1"
 msgstr "Word 1"
 
 msgid "word2"
@@ -1033,10 +1006,10 @@ msgstr ""
 msgid "word4"
 msgstr ""
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
-                actual_content = f.read()
+            actual_content = f.read()
         assert expected_content == actual_content
 
     @freeze_time("1994-11-11")
@@ -1049,27 +1022,7 @@ msgstr ""
         self.cmd.finalize_options()
         self.cmd.run()
 
-        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
-# Copyright (C) 1994 ORGANIZATION
-# This file is distributed under the same license as the PROJECT project.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PROJECT VERSION\n"
-"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: {date}\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel {VERSION}\n"
-
-msgid "word1"
+        expected_content = self._get_expected(fr"""msgid "word1"
 msgstr "Word 1"
 
 msgid "word2"
@@ -1081,44 +1034,24 @@ msgstr ""
 msgid "word4"
 msgstr "Word 4"
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
-                actual_content = f.read()
+            actual_content = f.read()
         assert expected_content == actual_content
 
     @freeze_time("1994-11-11")
-    def test_compenidum(self):
+    def test_compenidum_overwrite(self):
         self.cmd.input_files = [self.temp_def, self.temp_ref]
         self.cmd.output_file = self.output_file
         self.cmd.compendium = self.compendium
         self.cmd.no_fuzzy_matching = True
         self.cmd.no_compendium_comment = True
-        self.cmd.c_overwrite = True
+        self.cmd.compendium_overwrite = True
         self.cmd.finalize_options()
         self.cmd.run()
 
-        date = format_datetime(datetime(1994, 11, 11, 00, 00), 'yyyy-MM-dd HH:mmZ', tzinfo=LOCALTZ, locale='en')
-        expected_content = fr"""# Translations template for PROJECT.
-# Copyright (C) 1994 ORGANIZATION
-# This file is distributed under the same license as the PROJECT project.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 1994.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PROJECT VERSION\n"
-"Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: {date}\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel {VERSION}\n"
-
-msgid "word1"
+        expected_content = self._get_expected(fr"""msgid "word1"
 msgstr "Comp Word 1"
 
 msgid "word2"
@@ -1136,11 +1069,63 @@ msgstr "Word 4"
 #~ msgid "word2"
 #~ msgstr "Word 2"
 
-"""
+""")
 
         with open(self.output_file, 'r') as f:
-                actual_content = f.read()
+            actual_content = f.read()
         assert expected_content == actual_content
+
+    def test_update(self):
+        self.cmd.input_files = [self.temp_def, self.temp_ref]
+        self.cmd.update = True
+        self.cmd.no_fuzzy_matching = True
+        self.cmd.finalize_options()
+        self.cmd.run()
+
+        expected_content = self._get_expected(fr"""msgid "word1"
+msgstr "Word 1"
+
+msgid "word2"
+msgstr "Word 2"
+
+msgid "word3"
+msgstr ""
+
+msgid "word4"
+msgstr ""
+
+""")
+
+        with open(self.temp_def, 'r') as f:
+            actual_content = f.read()
+        assert expected_content == actual_content
+
+    def test_update_backup(self):
+        with open(self.temp_def, 'r') as f:
+            before_content = f.read()
+
+        self.cmd.input_files = [self.temp_def, self.temp_ref]
+        self.cmd.update = True
+        self.cmd.backup = True
+        self.cmd.no_fuzzy_matching = True
+        self.cmd.finalize_options()
+        self.cmd.run()
+
+        assert os.path.exists(self.temp_def + '~') 
+        with open(self.temp_def + '~', 'r') as f:
+            actual_content = f.read()
+        assert before_content == actual_content
+
+        os.unlink(self.temp_def)
+        shutil.move(self.temp_def + '~', self.temp_def)
+        self.cmd.suffix = '.bac'
+        self.cmd.run()
+
+        assert os.path.exists(self.temp_def + '.bac')
+        with open(self.temp_def + '.bac', 'r') as f:
+            actual_content = f.read()
+        assert before_content == actual_content
+
 
 class CommandLineInterfaceTestCase(unittest.TestCase):
